Title: 자연수 n개의 합, Time: 5ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240716/자연수 n개의 합/sum-of-n-natural-numbers.cpp
@@ -1,13 +1,13 @@
 #include <iostream>
 using namespace std;
 
-int s;
+long long s;
 
 int main() {
     cin >> s;
-    int n = 0;
+    long long n = 0;
 
-    int left = 1, right = s, mid = 0;
+    long long left = 1, right = s, mid = 0;
     while(left <= right) {
         mid = (left + right)/2;
